Fix transformation of <script> tags. Closes #84
The problem was that the transformScriptTags attached to DOMContentLoaded took one argument (the event), which it was treating as a list of script tags to transform. This is obviously wrong. DOM-specific stuff like this needs more unit testing in babel-standalone.

Tested scriptTag-src.html and scriptTag-custom.html examples, both work as expected